리눅스에서 컴파일되도록 수정함.

+++ trunk/SipClient/SipClient.cpp
@@ -86,7 +86,7 @@ void CSipClient::SipLog( bool bSend, const char * pszPacket )
 	localtime_r( &sttTime.tv_sec, &sttTm );	
 #endif
 
-	printf( "[%02d:%02d:%02d.%03d] %s [%s]\n", sttTm.tm_hour, sttTm.tm_min, sttTm.tm_sec, sttTime.tv_usec / 1000, bSend ? "Send" : "Recv", pszPacket );
+	printf( "[%02d:%02d:%02d.%03d] %s [%s]\n", sttTm.tm_hour, sttTm.tm_min, sttTm.tm_sec, (int)(sttTime.tv_usec / 1000), bSend ? "Send" : "Recv", pszPacket );
 }
 
 int main( int argc, char * argv[] )
